+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## 0.3.6.0 -- 2023-12-31
 * First public version
 
+## 0.3.8.1
+* Added getConnectionError to get any unrecoverable errors seen by the
+  thread that performs reconnections
+
 ## 0.3.8.0
 * Changed interface of 'SubscribeCmd' to allow passing a callback that will get
   called when the SUBSCRIBE and UNSUBSCRIBE requestes have been acknowledged by